feat: Improve target selection
When finding the last message from a trigger in the main stream, always pick a target which is also in the main stream.
When the relation is not explicit, we now grab the context of the event and paginate backwards until we find a message-like event which is not a part of a thread.